[69. Sqrt(x)][Compile Error]committed by LeetCode Extension

+++ 69-Sqrt(x)/solution.java
@@ -5,7 +5,7 @@ public class Solution {
         int left = 0, right = x;
         
         while(left<= right) {
-            long mid = left + (right - left)/2;
+            int mid = left + (right - left)/2;
             if(mid*mid > x) {
                 right = mid - 1;
             } else {
